refactor components and add stylings
