+++ src/api/apiKey.js
@@ -1,0 +1,3 @@
+const apiKey = () => process.env.REACT_APP_API_KEY;
+
+export default apiKey;
