+++ src/components/Navbar/Navbar.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
+import './Navbar.module.css';
+
 const Navbar = () => (
-  <header className="bg-primary">
-    <nav className="text-center">
+  <header className="bg-success text-center">
+    <nav>
       <Link to="/">
         <h1 className="text-white">Tasty Recipes</h1>
       </Link>
