+++ src/components/Recipe/Recipe.js
@@ -1,12 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import styles from './Recipe.module.css';
+
 const Recipe = ({ recipe }) => {
   const { title, image } = recipe;
 
   return (
-    <div>
-      <img src={image} alt={title} />
+    <div className={styles.recipeContainer}>
+      <div className={styles.imgContainer}>
+        <img src={image} alt={title} />
+      </div>
       <h3>{title}</h3>
     </div>
   );
