+++ src/components/RecipeFilter/RecipeFilter.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import './RecipeFilter.module.css';
+
 const RecipeFilter = ({ updateFilter, filter }) => {
   const categories = [
     'Chicken',
@@ -15,13 +17,11 @@ const RecipeFilter = ({ updateFilter, filter }) => {
   };
 
   return (
-    <div>
-      <select onChange={handleFilterChange} name="category" defaultValue={filter}>
-        {categories.map(category => (
-          <option key={category}>{category}</option>
-        ))}
-      </select>
-    </div>
+    <select onChange={handleFilterChange} name="category" defaultValue={filter}>
+      {categories.map(category => (
+        <option key={category}>{category}</option>
+      ))}
+    </select>
   );
 };
