+++ src/index.css
@@ -14,6 +14,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  color: #555;
 }
 
 code {
@@ -24,4 +25,8 @@ code {
     Consolas,
     'Courier New',
     monospace;
+}
+
+a:hover {
+  text-decoration: none;
 }
